pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "process-gpt-agent-sdk"
version = "0.2.0"
description = "Supabase 기반 이벤트/작업 폴링으로 A2A AgentExecutor를 실행하는 SDK"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["agent", "a2a", "supabase", "workflow", "sdk", "processgpt"]
classifiers = [
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3 :: Only",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
]
dependencies = [
	"supabase>=2.0.0",
	"python-dotenv>=1.0.0",
	"click>=8.0.0",
	"asyncio-mqtt>=0.13.0",
	"jsonschema>=4.0.0",
	"structlog>=23.0.0",
	"typing-extensions>=4.0.0",
	"python-dateutil>=2.8.0",
	"openai>=1.40.0",
	"a2a-sdk==0.2.4",
	"requests>=2.31.0",
	"anyio>=4.4.0",
	"pydantic>=2.7.0",
	"crewai>=0.51.0",
	"crewai-tools>=0.8.2",
	"mem0ai>=0.1.0",
	"mcp>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/process-gpt-agent-sdk"
Issues = "https://github.com/your-org/process-gpt-agent-sdk/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["processgpt_agent_sdk*"]

[tool.setuptools.package-data]
processgpt_agent_sdk = ["sql/*.sql"]
